add double hook to SeaCreatureFishEvent; add double hooks to shark fish counter
