+++ src/main/java/at/hannibal2/skyhanni/events/SeaCreatureFishEvent.kt
@@ -2,4 +2,8 @@ package at.hannibal2.skyhanni.events
 
 import at.hannibal2.skyhanni.features.fishing.SeaCreature
 
-class SeaCreatureFishEvent(val seaCreature: SeaCreature, val chatEvent: LorenzChatEvent) : LorenzEvent()
+class SeaCreatureFishEvent(
+    val seaCreature: SeaCreature,
+    val chatEvent: LorenzChatEvent,
+    val doubleHook: Boolean
+) : LorenzEvent()
