+++ src/main/java/at/hannibal2/skyhanni/features/fishing/SeaCreatureManager.kt
@@ -1,10 +1,32 @@
 package at.hannibal2.skyhanni.features.fishing
 
+import at.hannibal2.skyhanni.SkyHanniMod
+import at.hannibal2.skyhanni.events.LorenzChatEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
+import at.hannibal2.skyhanni.events.SeaCreatureFishEvent
 import at.hannibal2.skyhanni.utils.LorenzUtils
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 class SeaCreatureManager {
+
+    private var doubleHook = false
+
+    @SubscribeEvent
+    fun onChatMessage(event: LorenzChatEvent) {
+        if (!LorenzUtils.inSkyBlock) return
+        if (doubleHookMessages.contains(event.message)) {
+            if (SkyHanniMod.feature.fishing.compactDoubleHook) {
+                event.blockedReason = "double_hook"
+            }
+            doubleHook = true
+        } else {
+            val seaCreature = getSeaCreature(event.message)
+            if (seaCreature != null) {
+                SeaCreatureFishEvent(seaCreature, event, doubleHook).postAndCatch()
+            }
+            doubleHook = false
+        }
+    }
 
     @SubscribeEvent
     fun onRepoReload(event: RepositoryReloadEvent) {
@@ -43,6 +65,11 @@ class SeaCreatureManager {
     companion object {
         private val seaCreatureMap = mutableMapOf<String, SeaCreature>()
         var allFishingMobNames = emptyList<String>()
+
+        private val doubleHookMessages = setOf(
+            "§eIt's a §r§aDouble Hook§r§e! Woot woot!",
+            "§eIt's a §r§aDouble Hook§r§e!"
+        )
 
         fun getSeaCreature(message: String): SeaCreature? {
             return seaCreatureMap.getOrDefault(message, null)
