+++ src/main/java/at/hannibal2/skyhanni/features/fishing/SeaCreatureMessageShortener.kt
@@ -8,19 +8,7 @@ import net.minecraftforge.fml.common.eventhandler.EventPriority
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 class SeaCreatureMessageShortener {
-    private var nextIsDoubleHook: Boolean = false
-
     private val config get() = SkyHanniMod.feature.fishing
-
-    @SubscribeEvent(priority = EventPriority.LOW)
-    fun onChatMessage(event: LorenzChatEvent) {
-        if (!LorenzUtils.inSkyBlock) return
-        if (!config.shortenFishingMessage && !config.compactDoubleHook) return
-        if (doubleHookMessages.contains(event.message)) {
-            event.blockedReason = "double_hook"
-            nextIsDoubleHook = true
-        }
-    }
 
     @SubscribeEvent
     fun onSeaCreatureFish(event: SeaCreatureFishEvent) {
@@ -33,8 +21,7 @@ class SeaCreatureMessageShortener {
             "§9You caught a $seaCreature§9!"
         } else event.chatEvent.message
 
-        if (config.compactDoubleHook && nextIsDoubleHook) {
-            nextIsDoubleHook = false
+        if (config.compactDoubleHook && event.doubleHook) {
             message = "§e§lDOUBLE HOOK! $message"
         }
         LorenzUtils.chat(message)
@@ -42,12 +29,5 @@ class SeaCreatureMessageShortener {
         if (seaCreature.fishingExperience == 0) {
             LorenzUtils.debug("no fishing exp set for " + seaCreature.displayName)
         }
-    }
-
-    companion object {
-        private val doubleHookMessages = setOf(
-            "§eIt's a §r§aDouble Hook§r§e! Woot woot!",
-            "§eIt's a §r§aDouble Hook§r§e!"
-        )
     }
 }
